requirements.txt
numpy
tqdm
sklearn
opencv-python
prettytable
torchsummary
